Provide a way to opt-out under a buggy tmux session.
There appears to be an interaction between Konsole and tmux with
tmux 1.6.  In a very specific instance, doing a 'dd' under Vim can cause
the screen to shift and become slightly distorted.  It works fine with
tmux 1.7 though.  So provide users with a way to opt-out of changing the
cursor under tmux.