Only add the prefix to t_ti when under Konsole.
This fixes #19 where the prefix was causing issues with newer VTE-based
terminals.  Added a new variable, s:prefix_ti, instead of using cursor
escape style to make it clear it's tied to the terminal emulator and not
the style of escape sequences.

Thanks to Phong Cao for doing the leg work in tracking down the problem!