Allow disabling the t_ti initialization.
This is really to help workaround a bug seen in gnome-terminal (using
VTE 0.40.2) that was causing the cursor to disappear.  The user in #19
was running Arch Linux and was seeing this problem.  However, I was
unable to reproduce it under Fedora 22--which is also using VTE 0.40.2.

As a compromise, add a way for users to avoid `t_ti` initialization and
make a few notes about the issue and why this new option exists.