Change the default cursors to match gvim better.
We will now use a blinking line cursor in insert mode, if available.
And fallback to a line or blinking underline, depending on what's
supported by the terminal.  We will also default to a blinking block
cursor in normal mode, and default to just a plain block cursor when
leaving Vim.  Finally, the default replace mode cursor is now a blinking
underline.