Don't use ToggleCursorByMode() when vim has t_SR support
If Vim supports t_SR, then use that in preference to
ToggleCursorByMode().  Most of the scaffolding was already there, we
just had to disable the use of ToggleCursorByMode().

Changed the way `t_SR` support is detected by checking for `t_SR`
support directly, instead of checking the Vim version and patch level.